Implement configuration via Viper
Viper is the most widely used and safest way to manage configuration loading. Especially
if this becomes accepted as a tool for use within the company, we would want to allow for
configuration to be managed like this rather than with a hand-rolled alternative.

I've added the most widely used testing tool as well mostly to show how tests would be done in
idiomatic go.